BackEnd/requirements.txt
fastapi
pydantic
firebase-admin
google-generativeai
uvicorn